requirements.txt
akshare
mysql-connector-python
dbutils
PyYAML
pandas